Working updater
This build has a working updater.py module. I wrote it from scratch so there's little to no security, probably not ideal for a public release..

It will check it's internal "file_version" against the github .txt and pull the updated executable from gofile.io if needed. (I wanted to upload the executable to github but the file is too large(over 25mb).