Working Bot CLIs
This commit fixes OSBot and Tribot CLI integration. It also intelligently loads individual settings for each bot client when the user changes the combobox to select a different client.

Note: Paths are written to assume that there is a TRiBot folder in the src/ directory and the osbot jar file is also in the src/ directory.